Squashing: refactor OnlineIndexer to the structure below:
  OnlineIndexer: User API + defines OnlineIndexer.Config, OnlineIndexer.IndexFromIndexPolicy, OnlineIndexer.RecordBuiltRangeException
      Scanners:
      OnlineIndexerScanner: base class for scanners
          OnlineIndexerByRecords: scan by records (all records in DB)
          OnlineIndexerByIndex: scan by source index
             Scanners' Helper:
             OnlineIndexerThrottle: (implements buildAsync) wraps run/commit/retry with workload adjustment on success/failure
      Shared classes:
      OnlineIndexerCommon: holds the operation's setup for all classes
      OnlineIndexerException: scanner's exception

add test: non-VALUE source

Implement (some of) Scott's requests

Apply most of Tao's, Alec's, Scott's review requests.

spotbugsMain fixes

move indexBuildLockSubspace & indexBuildScannedRecordsSubspace to IndexingBase

renaming for clarity
