+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/logging/LogMessageKeys.java
@@ -146,6 +146,7 @@ public enum LogMessageKeys {
     INDEXER_CURR_RETRY("indexer_curr_retry"),
     INDEXER_MAX_RETRIES("indexer_max_retries"),
     DIRECTOY("directory"),
+    SOURCE_INDEX("source_index"),
 
     // comparisons
     COMPARISON_VALUE("comparison_value"),
